refactor date related functions into datetime.ts
