+++ src/components/Event/EventCard.vue
@@ -283,130 +283,7 @@ const to = computed(() => {
   return { name: RouteName.EVENT, params: { uuid: props.event.uuid } };
 });
 
-/*
-  Date formatting.
-
-  In the event card we want datetimes to be nicely formatted. The `formatedDate` contains whatever text
-  is to be displayed for that purpouse. The formatting is along the lines of
-    https://github.com/Kompismoln/klimatkalendern/issues/25
-  
-  TL;DR When the locale is sv we use custom formatting logic to have the date be super spick and span.
-  If the locale is something else we use generic l11n to make the datetime string.
-*/
-
-const formatedDate = computed(() => formatDate(props.event));
-
-function formatDate(event: IEvent) {
-  if (dateFnsLocale?.code === "sv") {
-    return formatDateSv(event);
-  }
-  return formateDateGeneric(event);
-}
-
-function formateDateGeneric(event: IEvent) {
-  if (!dateFnsLocale) {
-    console.error("Locale is undefined");
-    return;
-  }
-
-  const b = new Date(event.beginsOn);
-  const e = new Date(event.endsOn ?? b); // TODO: How can an event no have an end date?
-
-  // Single day event
-  if (
-    b.getFullYear() == e.getFullYear() &&
-    b.getMonth() == e.getMonth() &&
-    b.getDate() == e.getDate()
-  ) {
-    // No time information
-    if (!event.options.showStartTime) {
-      return `🗓 ${dtutils.formatDateForEvent(b, dateFnsLocale)}`;
-    }
-
-    // With start time.
-    if (event.options.showStartTime && !event.options.showEndTime) {
-      return `🗓 ${dtutils.formatDateTimeForEvent(b, dateFnsLocale)}`;
-    }
-
-    // Start and end time.
-    return `🗓 ${dtutils.formatTimeRangeForEvents(b, e, dateFnsLocale)}`;
-  }
-
-  // Multi day event
-  return `🗓 ${dtutils.formatDateForEvent(new Date(props.event.beginsOn), dateFnsLocale)} – ${dtutils.formatDateForEvent(e, dateFnsLocale)}`;
-}
-
-function formatDateSv(event: IEvent) {
-  const b = new Date(event.beginsOn);
-  const e = new Date(event.endsOn ?? b); // TODO: How can an event no have an end date?
-
-  const bWeekDay = dtutils.localeShortWeekDayNames(dateFnsLocale)[b.getDay()];
-  const eWeekDay = dtutils.localeShortWeekDayNames(dateFnsLocale)[e.getDay()];
-
-  // NOTE: We never write out which year the event occurs on,
-  //       I imagine the cases where that's important are excidingly rare.
-
-  // Single day event
-  if (
-    b.getFullYear() == e.getFullYear() &&
-    b.getMonth() == e.getMonth() &&
-    b.getDate() == e.getDate()
-  ) {
-    // No time information
-    if (!event.options.showStartTime) {
-      return `🗓 ${capitalize(bWeekDay)} ${b.getDate()} ${getMonthSv(b)}`;
-    }
-
-    // With start time.
-    if (event.options.showStartTime && !event.options.showEndTime) {
-      return `🗓 ${capitalize(bWeekDay)} ${b.getDate()} ${getMonthSv(b)} ${getTimeSv(b)}`;
-    }
-
-    // Start and end time.
-    return `🗓 ${capitalize(bWeekDay)} ${b.getDate()} ${getMonthSv(b)} ${getTimeRangeSv(b, e)}`;
-  }
-
-  // Multi day event
-  if (b.getFullYear() == e.getFullYear() && b.getMonth() == e.getMonth()) {
-    return `🗓 ${capitalize(bWeekDay)} ${b.getDate()} – ${eWeekDay} ${e.getDate()} ${getMonthSv(b)} ${getTimeRangeSv(b, e)}`;
-  }
-
-  // NOTE: This code path is taken for events spanning different years to! But
-  //       omitting year will not be confusing assuming events don't span many many monts
-  //       or even many years.
-  return `🗓 ${capitalize(bWeekDay)} ${b.getDate()} ${getMonthSv(e)} – ${eWeekDay} ${e.getDate()} ${getMonthSv(e)} ${getTimeRangeSv(b, e)}`;
-}
-
-function capitalize(s: string): string {
-  return s.charAt(0).toUpperCase() + s.slice(1);
-}
-
-function getMonthSv(date: Date) {
-  return [
-    "jan",
-    "feb",
-    "mars",
-    "apr",
-    "maj",
-    "jun",
-    "jul",
-    "aug",
-    "sep",
-    "oct",
-    "nov",
-    "dec",
-  ][date.getMonth()];
-}
-
-function getTimeRangeSv(fromTime: Date, toTime: Date) {
-  // TODO: Display only one time if `from` and `to` are the same
-  return `${getTimeSv(fromTime)}–${getTimeSv(toTime)}`;
-}
-
-function getTimeSv(date: Date) {
-  if (date.getMinutes() === 0) {
-    return `${date.getHours().toString().padStart(2, "0")}`;
-  }
-  return `${date.getHours().toString().padStart(2, "0")}.${date.getMinutes().toString().padStart(2, "0")}`;
-}
+const formatedDate = computed(() =>
+  dtutils.formatEventDatetime(props.event, dateFnsLocale)
+);
 </script>
